requirements.txt
asgiref==3.8.1
Django==4.2.2
sqlparse==0.5.1
tzdata==2024.1
python-dotenv
psycopg2-binary
pillow
black
ipython
pytils
django-crontab
redis